backend/requirements.txt
# HR Assistant RAG - Enterprise Edition (Phase 3) Dependencies
#
# Installation instructions:
# 1. Create virtual environment: python3 -m venv venv
# 2. Activate: source venv/bin/activate (macOS/Linux) or venv\Scripts\activate (Windows)
# 3. Install PyTorch first: pip install torch --index-url https://download.pytorch.org/whl/cpu
# 4. Install dependencies: pip install -r requirements.txt
# 5. Install huggingface_hub fix: pip install 'huggingface_hub<0.17'

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Database
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
alembic==1.13.0
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional, for production)

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.2

# ML/NLP (PyTorch must be installed separately - see above)
sentence-transformers==2.2.2
transformers==4.35.2
# torch==2.1.1  # Install separately: pip install torch --index-url https://download.pytorch.org/whl/cpu
# huggingface_hub<0.17  # Install after: pip install 'huggingface_hub<0.17'

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8
beautifulsoup4==4.12.2
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging & Monitoring
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
faker==22.0.0

# Development
black==23.12.1
flake8==7.0.0
